StockCounter - Created new module to contain methods, modified inventory validation check for better performance
