+++ src/main/java/com/Acrobot/ChestShop/ChestShop.java
@@ -17,11 +17,10 @@ import com.Acrobot.ChestShop.Listeners.Economy.ServerAccountCorrector;
 import com.Acrobot.ChestShop.Listeners.Economy.TaxModule;
 import com.Acrobot.ChestShop.Listeners.AuthMeChestShopListener;
 import com.Acrobot.ChestShop.Listeners.GarbageTextListener;
-import com.Acrobot.ChestShop.Listeners.Inventory.InventoryClose;
 import com.Acrobot.ChestShop.Listeners.Item.ItemMoveListener;
 import com.Acrobot.ChestShop.Listeners.ItemInfoListener;
 import com.Acrobot.ChestShop.Listeners.Modules.MetricsModule;
-import com.Acrobot.ChestShop.Listeners.PreShopCreation.StockCounterModifier;
+import com.Acrobot.ChestShop.Listeners.Modules.StockCounterModule;
 import com.Acrobot.ChestShop.Listeners.SignParseListener;
 import com.Acrobot.ChestShop.Listeners.Modules.DiscountModule;
 import com.Acrobot.ChestShop.Listeners.Modules.PriceRestrictionModule;
@@ -309,7 +308,6 @@ public class ChestShop extends JavaPlugin {
         registerPostShopCreationEvents();
         registerPostTransactionEvents();
         registerShopRemovalEvents();
-        registerInventoryEvents();
 
         registerModules();
 
@@ -382,14 +380,12 @@ public class ChestShop extends JavaPlugin {
         registerEvent(new ShopValidator());
         registerEvent(new SpamClickProtector());
         registerEvent(new StockFittingChecker());
-        registerEvent(new StockCounterModifier());
     }
 
     private void registerPostTransactionEvents() {
         registerEvent(new EconomicModule());
         registerEvent(new EmptyShopDeleter());
         registerEvent(new ItemManager());
-        registerEvent(new StockCounterManager());
         registerEvent(new TransactionLogger());
         registerEvent(new TransactionMessageSender());
     }
@@ -398,6 +394,7 @@ public class ChestShop extends JavaPlugin {
         registerEvent(new DiscountModule());
         registerEvent(new MetricsModule());
         registerEvent(new PriceRestrictionModule());
+        registerEvent(new StockCounterModule());
 
         registerEconomicalModules();
     }
@@ -405,10 +402,6 @@ public class ChestShop extends JavaPlugin {
     private void registerEconomicalModules() {
         registerEvent(new ServerAccountCorrector());
         registerEvent(new TaxModule());
-    }
-
-    private void registerInventoryEvents() {
-        registerEvent(new InventoryClose());
     }
 
     private void registerPluginMessagingChannels() {
